requirement.txt
alembic==0.8.4
amqp==1.4.9
anyjson==0.3.3
Babel==2.2.0
billiard==3.3.0.22
celery==3.1.19
coverage==4.1b1
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Alembic==1.2.1
Flask-BabelEx==0.9.2
Flask-Cache==0.13.1
Flask-Celery==2.4.3
Flask-Login==0.3.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.33
Mako==1.0.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==3.1.0
python-editor==0.5
pytz==2015.7
qrcode==5.1
requests==2.9.1
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.11
Werkzeug==0.11.3
WTForms==2.1
